created a formula to find the average individual tax paid for certain government departments

+++ src/main/kotlin/com/mshmelev/taxes/utils/CalculatedBudget.kt
@@ -1,4 +1,4 @@
 package com.mshmelev.taxes.utils
 
 data class CalculatedBudget(var department: String = "",
-                            var budgetAmount: Long = 0L)
+                            var budgetAmount: Double = 0.0)
